refactor: it seems that two of the four methods that were used can be merged into one IHS-201
