+++ doc_generation/content_gen_methods/__init__.py
@@ -2,13 +2,12 @@ from __future__ import annotations
 
 from doc_generation.content_gen_methods.command.command import ACommand
 from doc_generation.content_gen_methods.command.typer_command import TyperGroupCommand, TyperSingleCommand
+from doc_generation.content_gen_methods.mdx.mdx_code_doc import MdxCodeDocumentation
 
 from .base import ADocContentGenMethod
 from .command_output_method import CommandOutputDocContentGenMethod
 from .file_printing_method import FilePrintingDocContentGenMethod
 from .jinja2_method import Jinja2DocContentGenMethod
-from doc_generation.content_gen_methods.mdx.mdx_code_doc import MdxCodeDocumentation
-from .sdk_jinja2_method import SDKJinja2DocContentGenMethod
 
 __all__ = [
     "ACommand",
@@ -17,7 +16,6 @@ __all__ = [
     "FilePrintingDocContentGenMethod",
     "Jinja2DocContentGenMethod",
     "MdxCodeDocumentation",
-    "SDKJinja2DocContentGenMethod",
     "TyperGroupCommand",
     "TyperSingleCommand",
 ]
